+++ core/test/Dignite.Abp.Notifications.Tests/Dignite/Abp/Notifications/NotificationChannelRouting_Tests.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using Dignite.Abp.Notifications.Emailing;
 using Dignite.Abp.Notifications.SignalR;
 using Microsoft.AspNetCore.SignalR;
+using Microsoft.Extensions.Logging.Abstractions;
 using NSubstitute;
 using Shouldly;
 using Volo.Abp.EventBus.Distributed;
@@ -99,7 +100,17 @@ public class NotificationChannelRouting_Tests
         eventBus.WhenForAnyArgs(x => x.PublishAsync(Arg.Any<NotificationDeliveryEto>()))
             .Do(ci => published = ci.Arg<NotificationDeliveryEto>());
 
-        var distributor = new DefaultNotificationDistributor(store, definitionManager, eventBus);
+        var currentTenant = new TestCurrentTenant();
+        var distributor = new DefaultNotificationDistributor(
+            store,
+            definitionManager,
+            eventBus,
+            new DefaultNotificationRecipientEligibilityEvaluator(
+                definitionManager,
+                currentTenant,
+                NullLogger<DefaultNotificationRecipientEligibilityEvaluator>.Instance),
+            currentTenant,
+            NullLogger<DefaultNotificationDistributor>.Instance);
         await distributor.DistributeAsync(
             new NotificationInfo { Id = Guid.NewGuid(), NotificationName = "test" }, new[] { Guid.NewGuid() });